Create Kr hajong index.html
Q1. What is political theory? Discuss the significance of political theory.

Ans: At the most general level, political theory is 'a body of knowledge related to the phenomenon of the state'. While 'theory' refers to 'a systematic knowledge', 'political' refers to matters of public concern' According to David Held, political theory is a 'network of concepts and generalizations about political life involving ideas, assumptions and statements about the nature, purpose and key features of government, state and society, and about the political capabilities of human beings'l. Andrew Hacker defines it as 'a combination of a disinterested search for the principles of good state and good society on the one hand, and a disinterested search for knowledge of political and social reality on the other

NATURE AND SIGNIFICANCE

distinction between political theory and political thought, political philosophy and political science. As stated above, political theory is the study of the phenomena of the state both from philosophical as well as empirical points of view. In this context, certain similar terms are also used such as political thought, political philosophy, political science. Although all of them are concerned with explaining the political phenomena, yet political theory is distinct from them. The distinction of political theory from other terms is as follows, political theory and political thought It is generally believed that political thought is the general thought comprising of theories and values of all those persons or a section of the community who think and write on the day-do-day activities, policies and decisions of the state, and which has a 4 bearing on our present living. These persons can be philosophers, writers, journalists, poets, political commentators etc. Political thought has no 'fixed' form and can be in the form of treatise, speeches, political commentaries etc. What is important about political thought is that it is 'time bound' since the policies and programmes of the governments change from time to time. Thus we have Greek thought or Roman thought of ancient period or the political thought of the medieval ages. Political theory, on the other hand, is the systematic speculation of a particular writer who talks specifically about the phenomena of the state.

Political theory and political science As a discipline, political science is much more comprehensive and includes different forms of speculation in politics such as political thought, political theory, political philosophy, political- ideology, institutional or structural framework